print_schedule: Initial commit of markdown schedule formatting
Its not great, but after using it it required very little data
massaging to get it in place.

It might be possible in the future to go right from spreadsheet to
markdown a lot more complete. However it would require separating
out spreadsheet values from 'Time' to 'Day' 'Start' 'End' and
'Location'. With some fields being optional.